+++ examples/example__space_analysis__001.py
@@ -21,24 +21,17 @@ import sys
 
 sys.path.insert(0, os.path.abspath("..\\pycatia"))
 ##########################################################
+from pathlib import Path
 
 from pycatia import catia
-from pycatia.mec_mod_interfaces.part import Part
 from pycatia.mec_mod_interfaces.part_document import PartDocument
 
 # initialise the catia automation application
 caa = catia()
 documents = caa.documents
-documents.open(r"tests/cat_files/part_measurable.CATPart")
-
-document = PartDocument(caa.active_document.com_object)
-part = Part(document.part.com_object)
-# Note: It's not necessary to explicitly use the PartDocument or the Part class
-# with the com_object. It's perfectly fine to write it like this:
-#   document = caa.active_document
-#   part = document.part
-# But declaring 'document' and 'part' this way, your linter can't resolve the
-# product reference, see https://github.com/evereux/pycatia/issues/107#issuecomment-1336195688
+# if the active document is a CATPart this will return a PartDocument
+part_document: PartDocument = documents.open(Path(os.getcwd(), r"tests/cat_files/part_measurable.CATPart"))
+part = part_document.part
 
 # get the Bodies() collection
 bodies = part.bodies
@@ -54,7 +47,7 @@ body = bodies.item(1)
 # >>> Body(name="AnotherPartBody")
 
 # initialise the spa workbench
-spa_workbench = document.spa_workbench()
+spa_workbench = part_document.spa_workbench()
 # create a reference to measure.
 reference = part.create_reference_from_object(body)